+++ include/rogue/entities/Entity.h
@@ -4,23 +4,23 @@ namespace rogue {
 
 class Entity {
 public:
-  Entity(int x, int y, char symbol, int color, int maxHp);
+  Entity(float x, float y, char symbol, int color, int maxHp);
   virtual ~Entity() = default;
 
-  int getX() const;
-  int getY() const;
+  float getX() const;
+  float getY() const;
   char getSymbol() const;
   int getColor() const;
   int getHp() const;
   bool isDead() const;
   void takeDamage(int damage);
-  void setPosition(int newX, int newY);
+  void setPosition(float newX, float newY);
   virtual void update() = 0; // Pure virtual function for updating entity state
-  void move(int dx, int dy);
+  void move(float dx, float dy);
 
 protected:
-  int x;
-  int y;
+  float x;
+  float y;
   char symbol;
   int color;
   int hp;
