+++ include/rogue/entities/Player.h
@@ -5,8 +5,8 @@ namespace rogue {
 
 class Player : public Entity {
 public:
-  Player(int x, int y, char symbol, int color, int maxHp);
+  Player(float x, float y, char symbol, int color, int maxHp);
   void update() override; // Implement player-specific behavior
-  void move(int dx, int dy);
+  void move(float dx, float dy);
 };
 } // namespace rogue
